--- .eslintrc
@@ -1,3 +1,0 @@
-{
-  "parser": "babel-eslint"
-}
