+++ .gitignore
@@ -30,3 +30,5 @@ npm-debug.log*
 yarn-debug.log*
 
 yarn-error.log*
+
+.eslintrc
