--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-  "prettier.eslintIntegration": true,
-  "editor.formatOnSave": false,
-  "[javascript]": {
-    "editor.formatOnSave": false
-  }
-}
